netlify/functions/ask/requirements.txt
numpy==1.24.1
openai==0.26.1
pandas==1.5.2
python-dotenv==1.0.0
tiktoken==0.1.2
